Fixing Issue #17
Apparently it's possible for HttpServletRequest.getRequestURI() to return null.  Looking at Jetty 9 source specifically, there is a case in which the Servlet container could return null although the servlet spec seems to imply that returning null is invalid/impossible.  I've decided not to do anything about this and just ignore the fact that the servlet container could be wrong and have bugs, and therefore, it's not Curacao's job to work around those bugs by adding null checks everywhere.